Remove unecessary and broken code
The ContentNegotiation implementations were setting the Content-Type in
the createResponse method.  This is unnecessary because the middleware
functions were already setting the correct content type.  Besides being
uncessary, it also spelled Content-Type wrong and therefore had no
effect.  Note that it is possible for a user to pass a Content-Type
header and override this value, but this would be unwise.  I will
consider throwing or logging an error when this happens.